requirements.txt
pymongo==2.8
flask-mongoengine==0.7.1
pytest==2.7.1
flask-bcrypt==0.6.2
flask_kvsession==0.6.2
pytest-flask==0.8.1
webargs==0.14.0